+++ src/dispatching2/dialogue/mod.rs
@@ -92,8 +92,7 @@ pub use crate::dispatching::dialogue::{RedisStorage, RedisStorageError};
 pub use crate::dispatching::dialogue::{SqliteStorage, SqliteStorageError};
 
 pub use crate::dispatching::dialogue::{
-    serializer, ErasedStorage, InMemStorage, InMemStorageError, Serializer, Storage, StorageExt,
-    TraceStorage,
+    serializer, ErasedStorage, InMemStorage, InMemStorageError, Serializer, Storage, TraceStorage,
 };
 pub use get_chat_id::GetChatId;